refactor: move LoadMoreButton to UI library directory
- Relocated LoadMoreButton component to src/components/ui/
- Updated import path in ActivityFeed component
- Verified application builds and runs correctly
- Completes task T004
